Remove MultiWAN providers page, now configuration is made from NetworkAdapters page
