+++ root/usr/share/nethesis/NethServer/Help/en/NethServer_Module_Providers.rst
@@ -7,9 +7,6 @@ you must configure individual WAN connection
 (ISP) and establish usage policies (for example prefer a connection over another one).
 
 
-Configure
-=========
-
 Select provider's management policy: 
 
 * Balance to use all connections at the same time
@@ -36,29 +33,4 @@ Send mail notification on provider status change
 
      * From: mail sender address
      * To: mail recipient address
-
-
-Create or Edit
-==============
-
-Create or edit the configuration of providers. 
-
-Name 
-     A name to identify the connection (ISP). Max 5 characters. 
-
-Enabled/disabled 
-     Enable or disable the provider.
-
-Weight 
-     The "weight" of the connection. 
-     Traffic will be routed proportionally to the weight: higher weight means more traffic.
-     A provider with a weight of 100 will receive twice the traffic of one with weight 50. 
-     Please, assign weights accordingly to connection bandwidth.
-     When using active-backup mode, the weight determines the use of the line. 
-     If the first provider has weight 100 and the second has weight 50,
-     the traffic is always sent to the first provider. The second one will be used only if first provider goes down.
-
-Description 
-     An optional description to identify the provider. 
-
